Constructor injection

+++ src/main/java/com/origins/pgd/repository/masters/CourseRepository.java
@@ -7,11 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigInteger;
+
 
 /**
  * Created by Manoj Janaka on 14-11-2016.
  */
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, BigInteger> {
     @Query(value = "SELECT course FROM Course course WHERE course.code=:code")
     Course findOneByCode(@Param("code") String code);
