+++ src/main/java/com/origins/pgd/repository/transactions/CoursePaymentRepository.java
@@ -1,23 +1,17 @@
 package com.origins.pgd.repository.transactions;
 
-import com.origins.pgd.domain.dao.CoursePaymentDao;
-import com.origins.pgd.domain.masters.PaymentMode;
 import com.origins.pgd.domain.transactions.CoursePayment;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-import java.lang.annotation.Native;
+import java.math.BigInteger;
 import java.util.List;
-import java.util.Map;
 
 
 /**
  * Created by Manoj Janaka on 14-11-2016.
  */
-public interface CoursePaymentRepository extends JpaRepository<CoursePayment, Integer> {
+public interface CoursePaymentRepository extends JpaRepository<CoursePayment, BigInteger> {
 
 //    @Query(value = "SELECT P FROM CoursePayment P left join CourseWisePayments C ON C.id=P.coursePayment ")
 //    Page<CoursePayment> findAll(Pageable pageable);
